Criar permissões para `Equipamentos\Lista\*`

+++ app/Http/Controllers/Admin/Equipamentos/Lista/ListaController.php
@@ -9,12 +9,14 @@ use App\Models\Equipamentos\Cadastro\Equipamento;
 use App\Models\Equipamentos\Lista\Lista;
 use App\Models\Equipamentos\Lista\ProdutoLista;
 use Inertia\Inertia;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 
 class ListaController extends Controller
 {
     public function inicio()
     {
+        Gate::authorize('ver', Lista::class);
         $listas = Lista::paginate();
 
         return Inertia::render('Admin/Equipamentos/Lista/Inicio', compact('listas'));
@@ -22,11 +24,13 @@ class ListaController extends Controller
 
     public function criar()
     {
+        Gate::authorize('criar', Lista::class);
         return Inertia::render('Admin/Equipamentos/Lista/Criar');
     }
 
     public function salvar(ListaRequest $request)
     {
+        Gate::authorize('criar', Lista::class);
         $lista = new Lista($request->all());
         $lista->slug = ($lista->slug ?? Str::kebab($lista->nome));
         $lista->save();
@@ -37,6 +41,7 @@ class ListaController extends Controller
     public function editar(int $id)
     {
         $lista = Lista::findOrFail($id);
+        Gate::authorize('editar', Lista::class);
 
         return Inertia::render('Admin/Equipamentos/Lista/Editar', compact('lista'));
     }
@@ -44,6 +49,7 @@ class ListaController extends Controller
     public function atualizar(ListaRequest $request, int $id)
     {
         $lista = Lista::findOrFail($id);
+        Gate::authorize('editar', Lista::class);
         $lista->nome = $request->nome;
         $lista->slug = $request->slug ?? Str::kebab($request->nome);
         $lista->save();
@@ -53,7 +59,9 @@ class ListaController extends Controller
 
     public function excluir(int $id)
     {
-        Lista::findOrFail($id)->delete();
+        $lista = Lista::findOrFail($id);
+        Gate::authorize('excluir', Lista::class);
+        $lista->delete();
 
         return redirect()->route('admin.lista');
     }
@@ -61,6 +69,8 @@ class ListaController extends Controller
     public function adicionar(ProdutoListaRequest $request, int $listaId)
     {
         $lista = Lista::findOrFail($listaId);
+        Gate::authorize('adicionar', $lista);
+
         $equipamento = Equipamento::findOrFail($request->equipamento_id);
 
         ProdutoLista::firstOrCreate([
@@ -74,6 +84,8 @@ class ListaController extends Controller
     public function produtos(int $id)
     {
         $lista = Lista::findOrFail($id);
+        Gate::authorize('ver', $lista);
+
         $produtos = $lista->produtoLista()->with([
             'equipamento',
             'equipamento.modelo',
@@ -86,6 +98,9 @@ class ListaController extends Controller
 
     public function remover(int $listaId, int $produtoId)
     {
+        $lista = Lista::findOrFail($listaId);
+        Gate::authorize('remover', $lista);
+
         $produto = ProdutoLista::where('lista_id', $listaId)->findOrFail($produtoId);
         $produto->delete();
